Consolidated refresh commands. Use the config saved in the last run folder. Added polling to refresh.
