+++ test_runner/src/main/kotlin/ftl/cli/FirebaseCommand.kt
@@ -1,6 +1,7 @@
 package ftl.cli
 
 import ftl.cli.firebase.CancelCommand
+import ftl.cli.firebase.RefreshCommand
 import ftl.cli.firebase.TestCommand
 import ftl.cli.firebase.test.ios.IosDoctorCommand
 import picocli.CommandLine
@@ -11,6 +12,7 @@ import picocli.CommandLine.Command
     synopsisHeading = "",
     subcommands = [
         TestCommand::class,
+        RefreshCommand::class,
         CancelCommand::class,
         IosDoctorCommand::class
     ]
