+++ test_runner/src/main/kotlin/ftl/cli/firebase/test/AndroidCommand.kt
@@ -1,7 +1,6 @@
 package ftl.cli.firebase.test
 
 import ftl.cli.firebase.test.android.AndroidDoctorCommand
-import ftl.cli.firebase.test.android.AndroidRefreshCommand
 import ftl.cli.firebase.test.android.AndroidRunCommand
 import picocli.CommandLine
 import picocli.CommandLine.Command
@@ -11,7 +10,6 @@ import picocli.CommandLine.Command
     synopsisHeading = "",
     subcommands = [
         AndroidRunCommand::class,
-        AndroidRefreshCommand::class,
         AndroidDoctorCommand::class
     ]
 )
