+++ test_runner/src/main/kotlin/ftl/cli/firebase/test/IosCommand.kt
@@ -1,7 +1,6 @@
 package ftl.cli.firebase.test
 
 import ftl.cli.firebase.test.ios.IosDoctorCommand
-import ftl.cli.firebase.test.ios.IosRefreshCommand
 import ftl.cli.firebase.test.ios.IosRunCommand
 import picocli.CommandLine
 import picocli.CommandLine.Command
@@ -11,7 +10,6 @@ import picocli.CommandLine.Command
     synopsisHeading = "",
     subcommands = [
         IosRunCommand::class,
-        IosRefreshCommand::class,
         IosDoctorCommand::class
     ]
 )
