--- test_runner/src/main/kotlin/ftl/cli/firebase/test/android/AndroidRefreshCommand.kt
@@ -1,37 +1,0 @@
-package ftl.cli.firebase.test.android
-
-import ftl.args.AndroidArgs
-import ftl.config.FtlConstants
-import ftl.run.TestRunner
-import kotlinx.coroutines.experimental.runBlocking
-import picocli.CommandLine.Command
-import picocli.CommandLine.Option
-import java.nio.file.Paths
-
-@Command(
-    name = "refresh",
-    sortOptions = false,
-    headerHeading = "",
-    synopsisHeading = "%n",
-    descriptionHeading = "%n@|bold,underline Description:|@%n%n",
-    parameterListHeading = "%n@|bold,underline Parameters:|@%n",
-    optionListHeading = "%n@|bold,underline Options:|@%n",
-    header = ["Downloads results for the last Firebase Test Lab run"],
-    description = ["""Selects the most recent run in the results/ folder.
-Reads in the matrix_ids.json file. Refreshes any incomplete matrices.
-"""]
-)
-class AndroidRefreshCommand : Runnable {
-    override fun run() {
-        runBlocking {
-            val config = AndroidArgs.load(Paths.get(configPath))
-            TestRunner.refreshLastRun(config)
-        }
-    }
-
-    @Option(names = ["-c", "--config"], description = ["YAML config file path"])
-    var configPath: String = FtlConstants.defaultAndroidConfig
-
-    @Option(names = ["-h", "--help"], usageHelp = true, description = ["Prints this help message"])
-    var usageHelpRequested: Boolean = false
-}
