+++ test_runner/src/test/kotlin/ftl/cli/FirebaseCommandTest.kt
@@ -21,8 +21,9 @@ class FirebaseCommandTest {
             "firebase [COMMAND]\n" +
                 "Commands:\n" +
                 "  test\n" +
-                "  cancel  Cancels the last Firebase Test Lab run\n" +
-                "  doctor  Verifies flank firebase is setup correctly\n"
+                "  refresh  Downloads results for the last Firebase Test Lab run\n" +
+                "  cancel   Cancels the last Firebase Test Lab run\n" +
+                "  doctor   Verifies flank firebase is setup correctly\n"
         )
     }
 }
