+++ test_runner/src/test/kotlin/ftl/cli/firebase/RefreshCommandTest.kt
@@ -1,8 +1,8 @@
-package ftl.cli.firebase.test.ios
+package ftl.cli.firebase
+
 
 import com.google.common.truth.Truth
 import com.google.common.truth.Truth.assertThat
-import ftl.config.FtlConstants
 import ftl.test.util.FlankTestRunner
 import org.junit.Rule
 import org.junit.Test
@@ -11,52 +11,47 @@ import org.junit.runner.RunWith
 import picocli.CommandLine
 
 @RunWith(FlankTestRunner::class)
-class IosRefreshCommandTest {
+class RefreshCommandTest {
     @Rule
     @JvmField
     val systemOutRule: SystemOutRule = SystemOutRule().enableLog().muteForSuccessfulTests()
 
     @Test
-    fun iosRefreshCommandPrintsHelp() {
-        val refresh = IosRefreshCommand()
+    fun refreshCommandPrintsHelp() {
+        val refresh = RefreshCommand()
         assertThat(refresh.usageHelpRequested).isFalse()
         CommandLine.run<Runnable>(refresh, System.out, "-h")
 
         val output = systemOutRule.log
         Truth.assertThat(output).startsWith(
             "Downloads results for the last Firebase Test Lab run\n" +
-                "\n" +
-                "refresh [-h] [-c=<configPath>]\n" +
-                "\n" +
-                "Description:\n" +
-                "\n" +
-                "Selects the most recent run in the results/ folder.\n" +
-                "Reads in the matrix_ids.json file. Refreshes any incomplete matrices.\n" +
-                "\n" +
-                "\n" +
-                "Options:\n" +
-                "  -c, --config=<configPath>\n" +
-                "               YAML config file path\n" +
-                "  -h, --help   Prints this help message\n"
+                    "\n" +
+                    "refresh [-h]\n" +
+                    "\n" +
+                    "Description:\n" +
+                    "\n" +
+                    "Selects the most recent run in the results/ folder.\n" +
+                    "Reads in the matrix_ids.json file. Refreshes any incomplete matrices.\n" +
+                    "\n" +
+                    "\n" +
+                    "Options:\n" +
+                    "  -h, --help   Prints this help message\n"
         )
 
         assertThat(refresh.usageHelpRequested).isTrue()
     }
 
     @Test
-    fun iosRefreshCommandRuns() {
-        IosRefreshCommand().run()
+    fun refreshCommandRuns() {
+        val cmd = RefreshCommand()
+        cmd.run()
         val output = systemOutRule.log
         Truth.assertThat(output).contains("1 / 1 (100.00%)")
     }
 
     @Test
-    fun iosRefreshCommandOptions() {
-        val cmd = IosRefreshCommand()
-        assertThat(cmd.configPath).isEqualTo(FtlConstants.defaultIosConfig)
-        cmd.configPath = "tmp"
-        assertThat(cmd.configPath).isEqualTo("tmp")
-
+    fun refreshCommandOptions() {
+        val cmd = RefreshCommand()
         assertThat(cmd.usageHelpRequested).isFalse()
         cmd.usageHelpRequested = true
         assertThat(cmd.usageHelpRequested).isTrue()
